Rename one of the flow modules
