+++ tests/test_flowgraphs.py
@@ -8,7 +8,7 @@ from click.testing import CliRunner
 import perun.cli as cli
 import perun.profile.convert as convert
 import perun.logic.store as store
-import perun.view.flow.bokeh_factory as bokeh_graphs
+import perun.view.flow.factory as bokeh_graphs
 
 import tests.helpers.asserts as asserts
